+++ Installation/lib/cmake/CGAL/CGALConfig.cmake
@@ -164,6 +164,14 @@ foreach(cgal_lib ${CGAL_LIBRARIES})
   endif()
 endforeach()
 
+#
+# Define a specific target for basic viewer
+#
+add_library(CGAL::CGAL_Basic_viewer INTERFACE IMPORTED)
+  set_target_properties(CGAL::CGAL_Basic_viewer PROPERTIES
+    INTERFACE_COMPILE_DEFINITIONS "CGAL_USE_BASIC_VIEWER;QT_NO_KEYWORDS"
+    INTERFACE_LINK_LIBRARIES CGAL::CGAL_Qt5)
+
 include(${CGAL_CONFIG_DIR}/CGALConfigVersion.cmake)
 
 #